v2.5.1 (#1)
* Refactored unit tests

* No longer need to change the DateParseHandling setting for the Newtonsoft Json deserialization

The DateTime and DateTimeOffset specialized templates now handle the
case where the json token type is Date.

* Added benchmarks for the TemplateCompiler

micro-optimizations and disable DATAS to avoid skewing the performance
now that it is enabled by default in .NET 9.0

* Bug Fix: The EF Value converters were not properly supporting nullable fields

An invalid operation exception was thrown when the primitive had no
value, which happens when the table column is nullable and will have
no value.

* Code cleanup: using StringComparison enum in TypedPrimitiveAttribute  directly instead of object

* Conversion operator to underlying type is now implicit

* Typed primitives enhancements

- Typed primitives now implement IComparable<T> of the underlying
  primitive
- Added the Empty static readonly instance.
- Added the HasValue property.
- Added the GetValueOrDefault() and GetValueOrDefault(defauulValue)
  methods.
- CompareTo(object?) now supports comparisons with the underlying
  primitive type as well.

* Refactor typed primitives for improved null handling

This commit updates the handling of typed primitives by replacing the `IPrimitive` interface with `ITypedPrimitive`, which includes a `HasValue` property for better nullability checks. The `ToStringForJson` method in `Extensions.cs` was modified accordingly.

Test files, including `ReferencePrimitiveCreationTests.cs` and others, were updated to use `TTypedPrimitive` instead of `TPrimitive`, ensuring consistency in instance creation and value checking.

Additionally, the `Test` struct across various generated code files was enhanced with methods for safely retrieving wrapped values, and null checks were replaced with `HasValue` checks.

These changes aim to improve type safety, reliability, and clarity in the codebase.

* Refactor macro naming conventions for primitive types

Renamed the `Macros` class to `MacroNames` in `MacroNames.cs` and updated constants to follow a new naming scheme, such as `Namespace` to `TypedPrimitiveNamespace` and `TypeKeyword` to `PrimitiveName`.

Updated references in `TemplateProcessor.cs` to use the new `MacroNames` class.

Modified templates for `EFCoreValueConverter`, `NewtonsoftJsonConverter`, and `SystemTextJsonConverter` to align with the new naming conventions.

Adjusted `TypeManager.cs` to reflect the updated macro names, particularly for JSON converters and type attributes.

These changes improve code clarity and maintainability by standardizing naming conventions.

* Refactor value checks in converters

Updated to use HasValue instead of IsDefault, and GetValueOrDefault() instead of ValueOrDefault as these will be deprecated in a future version.

* Updated README to include all the changes in 2.5.1

---------

Co-authored-by: Eddie Velasquez <guardi@e-velasquez.com>